+++ src/MySQLReplication/Config/ConfigBuilder.php
@@ -72,6 +72,10 @@ class ConfigBuilder
      * @var array
      */
     private $custom = [];
+    /**
+     * @var int
+     */
+    private $heartbeatPeriod = 0;
 
     /**
      * @param string $user
@@ -231,18 +235,36 @@ class ConfigBuilder
 
     /**
      * @param int $tableCacheSize
+     * @return $this
      */
     public function withTableCacheSize($tableCacheSize)
     {
         $this->tableCacheSize = $tableCacheSize;
+
+        return $this;
     }
 
     /**
      * @param array $custom
+     * @return $this
      */
     public function withCustom(array $custom)
     {
         $this->custom = $custom;
+
+        return $this;
+    }
+
+    /**
+     * @see https://dev.mysql.com/doc/refman/5.6/en/change-master-to.html
+     * @param int $heartbeatPeriod
+     * @return $this
+     */
+    public function withHeartbeatPeriod($heartbeatPeriod)
+    {
+        $this->heartbeatPeriod = $heartbeatPeriod;
+
+        return $this;
     }
 
     /**
@@ -266,7 +288,8 @@ class ConfigBuilder
             $this->tablesOnly,
             $this->databasesOnly,
             $this->tableCacheSize,
-            $this->custom
+            $this->custom,
+            $this->heartbeatPeriod
         );
     }
 }