+++ src/MySQLReplication/Definitions/ConstEventsNames.php
@@ -18,4 +18,5 @@ class ConstEventsNames
     const WRITE = 'write';
     const MARIADB_GTID = 'mariadb gtid';
     const FORMAT_DESCRIPTION = 'format description';
+    const HEARTBEAT = 'heartbeat';
 }